Add workaround for screen being placed on external monitor when internal monitor exists.
This is to fix a bug I have with my laptop & an external HDMI screen. If I started a demo shell with the display connected, the primary QWindow would be placed on the external screen, not the built-in screen.

This is because when Qt asked for a Screen, it was returned the first screen Mir had in its list. Until we have utilities to configure displays, I've added basic prioritization of displays, so that internal displays are used first, then digitial, and finally analogue.