Enable use of MAppLauncherD
AppMan has to track the PID of the launch daemon. The launcher preallocates a mir connection for the app, resulting in a Session being registered and ready for the app. AppMan needs to track this session and associate it with the App using heuristics.

Note: the desktop_file_hint stuff won't work with this