Ensure RGB888 color format preferential to anything lower
For GL configs, if no color preference at all set by the user app, Qt's code ends up choosing RGB565 on android devices. This is the first config that EGL returns to us on Android devices, which is why is is chosen. However on desktop, it tends to be the highest color option.

But if you express a color preference, EGL is supposed to sort the list with the highest quality color at the top, and Qt does choose that correctly.

So to combat this quirk of EGL and Android, if user has set no color preference, default to high color values and try to get a suitable GL config. If that fails, fall back to the original request.